Revert "🚨 CRITICAL FIX: Restore desktop launch and unify desktop/terminal behavior"
