--- test_launch_consistency.py
@@ -1,359 +1,0 @@
-#!/usr/bin/env python3
-"""
-Comprehensive test to verify that desktop and terminal launch methods
-produce IDENTICAL behavior in all aspects.
-"""
-
-import sys
-import os
-import subprocess
-import tempfile
-import time
-
-# Add current directory to Python path
-sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
-
-def capture_launch_behavior(command):
-    """Capture the behavior of a launch command"""
-    try:
-        # Set up environment to capture behavior without GUI
-        env = os.environ.copy()
-        env['QT_QPA_PLATFORM'] = 'offscreen'
-        
-        # Run the command and capture output
-        result = subprocess.run(
-            command,
-            capture_output=True,
-            text=True,
-            timeout=10,
-            env=env,
-            cwd=os.path.dirname(os.path.abspath(__file__))
-        )
-        
-        return {
-            'stdout': result.stdout,
-            'stderr': result.stderr,
-            'returncode': result.returncode,
-            'success': result.returncode == 0
-        }
-    except subprocess.TimeoutExpired:
-        return {
-            'stdout': '',
-            'stderr': 'Timeout expired',
-            'returncode': -1,
-            'success': False
-        }
-    except Exception as e:
-        return {
-            'stdout': '',
-            'stderr': str(e),
-            'returncode': -1,
-            'success': False
-        }
-
-def test_identical_launch_behavior():
-    """Test that main.py and bin/applergui behave identically"""
-    print("🧪 Testing Identical Launch Behavior...")
-    
-    # Test script that exercises both launch methods
-    test_script = '''
-import sys
-import os
-sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
-
-os.environ['QT_QPA_PLATFORM'] = 'offscreen'
-
-try:
-    from shared_launcher import unified_main, setup_environment, AppleTVRemoteGUI
-    from PyQt5.QtWidgets import QApplication
-    
-    print("=== LAUNCH TEST START ===")
-    
-    # Test environment setup
-    setup_environment()
-    print(f"PYTHONPATH: {os.environ.get('PYTHONPATH', 'Not set')}")
-    print(f"QT_QPA_PLATFORM: {os.environ.get('QT_QPA_PLATFORM', 'Not set')}")
-    print(f"DISPLAY: {os.environ.get('DISPLAY', 'Not set')}")
-    print(f"Working Directory: {os.getcwd()}")
-    
-    # Test GUI creation
-    app = QApplication(sys.argv)
-    window = AppleTVRemoteGUI()
-    
-    print(f"Window Title: {window.windowTitle()}")
-    print(f"Window Size: {window.size().width()}x{window.size().height()}")
-    print(f"Discovery Button: {window.discover_btn.text()}")
-    print(f"Buttons Count: {len([attr for attr in dir(window) if attr.endswith('_btn')])}")
-    
-    # Test styling
-    style = window.styleSheet()
-    print(f"Has Black Background: {'background-color: #000000' in style}")
-    print(f"Has White Text: {'color: #ffffff' in style}")
-    
-    print("=== LAUNCH TEST END ===")
-    app.quit()
-    
-except Exception as e:
-    print(f"ERROR: {e}")
-    import traceback
-    traceback.print_exc()
-    sys.exit(1)
-'''
-    
-    # Write test script to temporary file
-    with tempfile.NamedTemporaryFile(mode='w', suffix='.py', delete=False) as f:
-        f.write(test_script)
-        test_script_path = f.name
-    
-    try:
-        # Test main.py behavior
-        print("  Testing main.py behavior...")
-        main_result = capture_launch_behavior([
-            sys.executable, '-c', 
-            f'exec(open("{test_script_path}").read())'
-        ])
-        
-        # Test bin/applergui behavior  
-        print("  Testing bin/applergui behavior...")
-        applergui_result = capture_launch_behavior([
-            sys.executable, 'bin/applergui', '--test'  
-        ])
-        
-        # For the applergui test, we need to modify it to run our test script
-        # Let's just test the import behavior since both call unified_main
-        applergui_import_test = '''
-import sys
-import os
-script_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-sys.path.insert(0, script_dir)
-
-from shared_launcher import unified_main
-print("=== APPLERGUI IMPORT TEST ===")
-print("unified_main function imported successfully")
-print("=== APPLERGUI IMPORT TEST END ===")
-'''
-        
-        with tempfile.NamedTemporaryFile(mode='w', suffix='.py', delete=False) as f:
-            f.write(applergui_import_test)
-            applergui_test_path = f.name
-        
-        applergui_result = capture_launch_behavior([
-            sys.executable, applergui_test_path
-        ])
-        
-        # Analyze results
-        print("📊 Launch Behavior Analysis:")
-        print(f"  Main.py success: {main_result['success']}")
-        print(f"  Applergui success: {applergui_result['success']}")
-        
-        if main_result['success']:
-            main_output = main_result['stdout']
-            if "LAUNCH TEST START" in main_output and "LAUNCH TEST END" in main_output:
-                print("  ✅ Main.py test completed successfully")
-                
-                # Extract key values
-                lines = main_output.split('\n')
-                main_values = {}
-                for line in lines:
-                    if ': ' in line and not line.startswith('==='):
-                        key, value = line.split(': ', 1)
-                        main_values[key] = value
-                
-                print(f"  📋 Main.py values: {len(main_values)} captured")
-                return True
-            else:
-                print("  ❌ Main.py test incomplete")
-                return False
-        else:
-            print(f"  ❌ Main.py failed: {main_result['stderr']}")
-            return False
-        
-        if applergui_result['success']:
-            print("  ✅ Applergui import test successful")
-        else:
-            print(f"  ❌ Applergui import failed: {applergui_result['stderr']}")
-        
-        return main_result['success'] and applergui_result['success']
-        
-    finally:
-        # Clean up temporary files
-        try:
-            os.unlink(test_script_path)
-            os.unlink(applergui_test_path)
-        except:
-            pass
-
-def test_import_consistency():
-    """Test that both entry points import the same modules"""
-    print("🧪 Testing Import Consistency...")
-    
-    try:
-        # Test main.py imports
-        with open('main.py', 'r') as f:
-            main_content = f.read()
-        
-        # Test bin/applergui imports  
-        with open('bin/applergui', 'r') as f:
-            applergui_content = f.read()
-        
-        # Both should import unified_main from shared_launcher
-        main_imports_unified = 'from shared_launcher import unified_main' in main_content
-        applergui_imports_unified = 'from shared_launcher import unified_main' in applergui_content
-        
-        print(f"  Main.py imports unified_main: {main_imports_unified}")
-        print(f"  Applergui imports unified_main: {applergui_imports_unified}")
-        
-        # Both should call unified_main
-        main_calls_unified = 'unified_main()' in main_content
-        applergui_calls_unified = 'unified_main()' in applergui_content
-        
-        print(f"  Main.py calls unified_main: {main_calls_unified}")
-        print(f"  Applergui calls unified_main: {applergui_calls_unified}")
-        
-        success = all([
-            main_imports_unified,
-            applergui_imports_unified, 
-            main_calls_unified,
-            applergui_calls_unified
-        ])
-        
-        if success:
-            print("  ✅ Both entry points use identical import structure")
-        else:
-            print("  ❌ Entry points have different import structures")
-        
-        return success
-        
-    except Exception as e:
-        print(f"  ❌ Import consistency test failed: {e}")
-        return False
-
-def test_environment_setup():
-    """Test that environment setup produces identical results"""
-    print("🧪 Testing Environment Setup Consistency...")
-    
-    try:
-        from shared_launcher import setup_environment
-        
-        # Save original state
-        original_env = {
-            'PYTHONPATH': os.environ.get('PYTHONPATH', ''),
-            'QT_QPA_PLATFORM': os.environ.get('QT_QPA_PLATFORM', ''),
-            'DISPLAY': os.environ.get('DISPLAY', ''),
-            'cwd': os.getcwd()
-        }
-        
-        # Test environment setup multiple times
-        for i in range(3):
-            setup_environment()
-            
-            current_env = {
-                'PYTHONPATH': os.environ.get('PYTHONPATH', ''),
-                'QT_QPA_PLATFORM': os.environ.get('QT_QPA_PLATFORM', ''),
-                'DISPLAY': os.environ.get('DISPLAY', ''),
-                'cwd': os.getcwd()
-            }
-            
-            print(f"  Test {i+1}: PYTHONPATH={current_env['PYTHONPATH'][:50]}...")
-            print(f"  Test {i+1}: QT_QPA_PLATFORM={current_env['QT_QPA_PLATFORM']}")
-            print(f"  Test {i+1}: DISPLAY={current_env['DISPLAY']}")
-        
-        print("  ✅ Environment setup is consistent across multiple calls")
-        return True
-        
-    except Exception as e:
-        print(f"  ❌ Environment setup test failed: {e}")
-        return False
-
-def test_gui_component_consistency():
-    """Test that GUI components are identical regardless of launch method"""
-    print("🧪 Testing GUI Component Consistency...")
-    
-    try:
-        from shared_launcher import AppleTVRemoteGUI
-        from PyQt5.QtWidgets import QApplication
-        import sys
-        
-        # Set up offscreen environment
-        os.environ['QT_QPA_PLATFORM'] = 'offscreen'
-        app = QApplication(sys.argv)
-        
-        # Create multiple instances to test consistency
-        instances = []
-        for i in range(3):
-            window = AppleTVRemoteGUI()
-            instances.append({
-                'title': window.windowTitle(),
-                'size': f"{window.size().width()}x{window.size().height()}",
-                'discover_btn': window.discover_btn.text(),
-                'button_count': len([attr for attr in dir(window) if attr.endswith('_btn')]),
-                'has_volume_container': hasattr(window, 'volume_container'),
-                'style_has_black': 'background-color: #000000' in window.styleSheet()
-            })
-        
-        # Check all instances are identical
-        first_instance = instances[0]
-        all_identical = all(instance == first_instance for instance in instances)
-        
-        print(f"  Created {len(instances)} GUI instances")
-        print(f"  Window title: {first_instance['title']}")
-        print(f"  Window size: {first_instance['size']}")
-        print(f"  Button count: {first_instance['button_count']}")
-        print(f"  Has volume container: {first_instance['has_volume_container']}")
-        print(f"  Has black styling: {first_instance['style_has_black']}")
-        print(f"  All instances identical: {all_identical}")
-        
-        app.quit()
-        
-        if all_identical:
-            print("  ✅ GUI components are consistent across instances")
-        else:
-            print("  ❌ GUI components vary between instances")
-        
-        return all_identical
-        
-    except Exception as e:
-        print(f"  ❌ GUI consistency test failed: {e}")
-        return False
-
-def main():
-    """Run comprehensive launch consistency tests"""
-    print("🚀 Comprehensive Launch Consistency Validation")
-    print("=" * 60)
-    
-    tests = [
-        test_import_consistency,
-        test_environment_setup,
-        test_gui_component_consistency,
-        test_identical_launch_behavior
-    ]
-    
-    passed = 0
-    total = len(tests)
-    
-    for test in tests:
-        try:
-            if test():
-                passed += 1
-                print("✅ PASSED\n")
-            else:
-                print("❌ FAILED\n")
-        except Exception as e:
-            print(f"❌ FAILED with exception: {e}\n")
-    
-    print("=" * 60)
-    print(f"Final Results: {passed}/{total} consistency tests passed")
-    
-    if passed == total:
-        print("🎉 ALL CONSISTENCY TESTS PASSED!")
-        print("✅ Desktop and terminal launch methods are IDENTICAL")
-        print("✅ All critical fixes have been successfully implemented")
-        return True
-    else:
-        print("💥 Some consistency tests failed")
-        print("❌ Launch methods may not be identical")
-        return False
-
-if __name__ == "__main__":
-    success = main()
-    sys.exit(0 if success else 1)
