--- validation_final.py
@@ -1,163 +1,0 @@
-#!/usr/bin/env python3
-"""
-Final validation of critical fixes implementation
-"""
-
-import sys
-import os
-
-# Add current directory to Python path
-sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
-
-def main():
-    """Final validation of all critical fixes"""
-    
-    print("🍎 Apple TV Remote GUI - CRITICAL FIXES VALIDATION")
-    print("=" * 70)
-    print("")
-    
-    tests_passed = 0
-    total_tests = 4
-    
-    # Test 1: Unified Launch System
-    print("🔍 Test 1: Unified Launch System")
-    try:
-        from shared_launcher import unified_main, setup_environment, AppleTVRemoteGUI, DiscoveryWorker
-        
-        # Check main.py uses shared launcher
-        with open('main.py', 'r') as f:
-            main_content = f.read()
-        main_uses_shared = 'from shared_launcher import unified_main' in main_content
-        
-        # Check bin/applergui uses shared launcher
-        with open('bin/applergui', 'r') as f:
-            applergui_content = f.read()
-        applergui_uses_shared = 'from shared_launcher import unified_main' in applergui_content
-        
-        if main_uses_shared and applergui_uses_shared:
-            print("   ✅ Both main.py and bin/applergui use unified_main()")
-            print("   ✅ Launch consistency achieved")
-            tests_passed += 1
-        else:
-            print("   ❌ Launch methods are not consistent")
-            
-    except Exception as e:
-        print(f"   ❌ Test failed: {e}")
-    
-    print("")
-    
-    # Test 2: Environment Setup
-    print("🔍 Test 2: Environment Setup")
-    try:
-        setup_environment()
-        
-        vars_set = 0
-        required_vars = ['PYTHONPATH', 'QT_QPA_PLATFORM', 'DISPLAY']
-        for var in required_vars:
-            if var in os.environ:
-                vars_set += 1
-        
-        if vars_set == len(required_vars):
-            print("   ✅ All environment variables configured")
-            print("   ✅ Environment consistency achieved")
-            tests_passed += 1
-        else:
-            print(f"   ❌ Only {vars_set}/{len(required_vars)} environment variables set")
-            
-    except Exception as e:
-        print(f"   ❌ Test failed: {e}")
-    
-    print("")
-    
-    # Test 3: GUI Components & Black Styling
-    print("🔍 Test 3: GUI Components & Black Styling")
-    try:
-        from PyQt5.QtWidgets import QApplication
-        
-        os.environ['QT_QPA_PLATFORM'] = 'offscreen'
-        app = QApplication(sys.argv)
-        
-        window = AppleTVRemoteGUI()
-        
-        # Check styling
-        style = window.styleSheet()
-        has_black_bg = 'background-color: #000000' in style
-        has_white_text = 'color: #ffffff' in style
-        has_volume = hasattr(window, 'volume_container')
-        
-        # Check buttons
-        button_count = len([attr for attr in dir(window) if attr.endswith('_btn')])
-        
-        if has_black_bg and has_white_text and has_volume and button_count >= 10:
-            print("   ✅ Black styling applied correctly")
-            print("   ✅ All GUI components present")
-            tests_passed += 1
-        else:
-            print("   ❌ GUI components or styling issues")
-        
-        app.quit()
-        
-    except Exception as e:
-        print(f"   ❌ Test failed: {e}")
-    
-    print("")
-    
-    # Test 4: Discovery System  
-    print("🔍 Test 4: Discovery System")
-    try:
-        # Test worker
-        worker = DiscoveryWorker()
-        has_signals = hasattr(worker, 'finished') and hasattr(worker, 'error')
-        
-        # Test parsing
-        mock_output = "Name: Test Apple TV\nAddress: 192.168.1.100\n----\nName: Test 2\nAddress: 192.168.1.101"
-        
-        app2 = QApplication([])
-        window2 = AppleTVRemoteGUI()
-        devices = window2._parse_atvremote_scan_output(mock_output)
-        app2.quit()
-        
-        if has_signals and len(devices) == 2:
-            print("   ✅ Discovery worker has proper signals")
-            print("   ✅ Discovery parsing works correctly")
-            tests_passed += 1
-        else:
-            print("   ❌ Discovery system issues")
-            
-    except Exception as e:
-        print(f"   ❌ Test failed: {e}")
-    
-    print("")
-    
-    # Results
-    print("🏆 VALIDATION RESULTS")
-    print("=" * 70)
-    print(f"Tests Passed: {tests_passed}/{total_tests}")
-    
-    if tests_passed == total_tests:
-        print("")
-        print("🎉 ALL CRITICAL FIXES VALIDATED SUCCESSFULLY!")
-        print("")
-        print("✅ Issues Fixed:")
-        print("   • Desktop App Completely Broken → FIXED")
-        print("   • Discovery Button Broken → FIXED")  
-        print("   • Frame White in Terminal → FIXED")
-        print("   • Launch Processes Different → FIXED")
-        print("")
-        print("🚀 ApplerGUI is now working with:")
-        print("   • Unified launch system (desktop & terminal identical)")
-        print("   • Fixed discovery button with proper threading")
-        print("   • Consistent black styling across all launch methods")
-        print("   • Robust error handling throughout")
-        print("")
-        print("💯 CRITICAL EMERGENCY FIXES COMPLETE!")
-        return True
-    else:
-        print("")
-        print("❌ Some critical fixes validation failed")
-        print("💥 Manual verification may be required")
-        return False
-
-if __name__ == "__main__":
-    success = main()
-    sys.exit(0 if success else 1)
